Separate Storage logic and provide option for custom storage
